Initial refactoring of common queue/topic functionality
